Rename RLE90 stuff to avoid collisions
