+++ utilities/compression/rle90_test.go
@@ -1,4 +1,4 @@
-package utilities
+package compression
 
 import (
 	"bytes"
